# api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/user/register":{"put":{"tags":["user-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}}}}},"/user/login":{"post":{"tags":["user-controller"],"operationId":"login","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultString"}}}}}}},"/tag/{id}":{"get":{"tags":["tag-controller"],"operationId":"getTagsByMemeId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultListTag"}}}}}},"post":{"tags":["tag-controller"],"operationId":"setTag","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}}}}},"/common/upload":{"post":{"tags":["common-controller"],"operationId":"upload","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultString"}}}}}}},"/user":{"get":{"tags":["user-controller"],"operationId":"getUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultString"}}}}}}},"/meme":{"get":{"tags":["meme-controller"],"operationId":"getUserMemeByPage","parameters":[{"name":"arg0","in":"query","required":true,"schema":{"$ref":"#/components/schemas/MemePageQueryDTO"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultListMemeVO"}}}}}}},"/meme/{id}":{"get":{"tags":["meme-controller"],"operationId":"getMemeById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultMemeVO"}}}}}}}},"components":{"schemas":{"UserDTO":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"Result":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"type":"object"}}},"ResultString":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"type":"string"}}},"ResultListTag":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/Tag"}}}},"Tag":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"}}},"MemePageQueryDTO":{"type":"object","properties":{"page":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"userId":{"type":"integer","format":"int32"}}},"MemeVO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"url":{"type":"string"},"userId":{"type":"integer","format":"int32"},"username":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"tags":{"type":"array","items":{"$ref":"#/components/schemas/Tag"}}}},"ResultListMemeVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/MemeVO"}}}},"ResultMemeVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"$ref":"#/components/schemas/MemeVO"}}}}}}